requirements.txt
argparse
selenium==3.141.0
numpy
easy_dict
pyyaml